sahayak-agent-backend/requirements.txt
functions-framework
google-cloud-aiplatform
google-cloud-firestore
google-api-python-client
google-auth-httplib2
google-auth-oauthlib